Revert "POO | java.lang.Integer"
This reverts commit 8336adb6d397adb814c6864b8c427591f4d1e447.

+++ src/java/class-sample/CE3104-POO-HernandezValerie/Geometry/src/ce3104/poo/geometry/shapes/Circle.java
@@ -2,9 +2,9 @@ package ce3104.poo.geometry.shapes;
 
 public final class Circle extends Shape {
 
-    private final Integer radius;
+    private final Float radius;
 
-    public Circle(Integer radius) {
+    public Circle(Float radius) {
         this.radius = radius;
     }
