+++ src/java/class-sample/CE3104-POO-HernandezValerie/Geometry/src/ce3104/poo/geometry/shapes/Rectangle.java
@@ -2,21 +2,21 @@ package ce3104.poo.geometry.shapes;
 
 public final class Rectangle extends Shape {
 
-    private final Integer length;
-    private final Integer width;
+    private final Float length;
+    private final Float width;
 
-    public Rectangle(Integer length, Integer width) {
+    public Rectangle(Float length, Float width) {
         this.length = length;
         this.width = width;
     }
 
     @Override
     public Float getPerimeter() {
-        return (float) (2 * (this.length + this.width));
+        return (2 * (this.length + this.width));
     }
 
     @Override
     public Float getArea() {
-        return (float) (this.length * this.width);
+        return (this.length * this.width);
     }
 }
